requirements.txt
pandas>=2.0
pyyaml>=6.0
feedparser>=6.0
requests>=2.32
